# frontend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pomodoro": {
      "post": {
        "operationId": "pomodoroCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PomodoroConfigurationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pomodoro configuration created.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PomodoroConfigurationResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a Pomodoro configuration",
        "tags": [
          "Pomodoro"
        ]
      },
      "get": {
        "operationId": "pomodoroList",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pomodoro configurations returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PomodoroConfigurationResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List my Pomodoro configurations",
        "tags": [
          "Pomodoro"
        ]
      }
    },
    "/pomodoro/{id}": {
      "get": {
        "operationId": "pomodoroGet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomodoro configuration returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PomodoroConfigurationResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one Pomodoro configuration",
        "tags": [
          "Pomodoro"
        ]
      },
      "put": {
        "operationId": "pomodoroUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PomodoroConfigurationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pomodoro configuration updated.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PomodoroConfigurationResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a Pomodoro configuration",
        "tags": [
          "Pomodoro"
        ]
      },
      "delete": {
        "operationId": "pomodoroDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomodoro configuration deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DeletePomodoroResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete configuration without deleting history",
        "tags": [
          "Pomodoro"
        ]
      }
    },
...
    "/youtube/videos": {
      "get": {
        "operationId": "youtubeVideoSearch",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "example": "lofi study",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "YouTube videos returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MediaItemResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Search public YouTube videos",
        "tags": [
          "YouTube videos"
        ]
      }
    },
    "/youtube/videos/resolve": {
      "post": {
        "operationId": "youtubeVideoResolve",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveYoutubeVideoRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "YouTube video metadata returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MediaItemResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Resolve a YouTube video URL",
        "tags": [
          "YouTube videos"
        ]
      }
    }
  },
  "info": {
    "title": "Pomodoro and Music API",
    "description": "Backend API for Pomodoro sessions and music playback.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "name"
          },
          "message": {
            "type": "string",
            "example": "Name is required"
          }
        },
        "required": [
          "message"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Invalid request"
          },
          "error_code": {
            "type": "string",
            "example": "INVALID_INPUT"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            }
          },
          "request_id": {
            "type": "string",
            "example": "0198bb70-92ac-77c8-b22f-892d2d97f599"
          }
        },
        "required": [
          "code",
          "message",
          "error_code",
          "details",
          "request_id"
        ]
      },
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "object",
            "example": "success"
          },
          "code": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "status",
          "code",
          "message",
          "data"
        ]
      },
...
      "PomodoroConfigurationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "focusDurationSeconds": {
            "type": "number"
          },
          "shortBreakDurationSeconds": {
            "type": "number"
          },
          "longBreakDurationSeconds": {
            "type": "number"
          },
          "focusSessionsBeforeLongBreak": {
            "type": "number"
          },
          "focusPlaylistId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "breakPlaylistId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "focusDurationSeconds",
          "shortBreakDurationSeconds",
          "longBreakDurationSeconds",
          "focusSessionsBeforeLongBreak",
          "focusPlaylistId",
          "breakPlaylistId",
          "isDefault",
          "createdAt",
          "updatedAt"
        ]
      },
      "PomodoroConfigurationRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Deep work"
          },
          "focusDurationSeconds": {
            "type": "number",
            "example": 1500
          },
          "shortBreakDurationSeconds": {
            "type": "number",
            "example": 300
          },
          "longBreakDurationSeconds": {
            "type": "number",
            "example": 900
          },
          "focusSessionsBeforeLongBreak": {
            "type": "number",
            "example": 4
          },
          "focusPlaylistId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "breakPlaylistId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "name",
          "focusDurationSeconds",
          "shortBreakDurationSeconds",
          "longBreakDurationSeconds",
          "focusSessionsBeforeLongBreak"
        ]
      },
      "DeletePomodoroResponseDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "MediaItemResponseDto": {
        "type": "object",
        "properties": {
          "externalMediaId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "channelName": {
            "type": "string",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "durationSeconds": {
            "type": "number",
            "nullable": true
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri"
          },
          "availability": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE",
              "PRIVATE",
              "DELETED",
              "REGION_BLOCKED",
              "UNKNOWN"
            ]
          }
        },
        "required": [
          "externalMediaId",
          "title",
          "channelName",
          "thumbnailUrl",
          "durationSeconds",
          "sourceUrl",
          "availability"
        ]
      },
...
      "ResolveYoutubeVideoRequestDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://www.youtube.com/watch?v=jfKfPfyJRdk"
          }
        },
        "required": [
          "url"
        ]
      }
    }
  }
}